Improve flexibility of collections
- Allow callbacks passed to `ICollection` methods to receive item keys,
  values or both
- Remove `$value` parameter from `ICollection::unset()`
- Move parts of `TCollection` to a separate `TReadableCollection` trait
